chore: rename table file
