+++ front/src/addons/events/pages/EventsList.tsx
@@ -5,7 +5,7 @@ import { Link } from 'react-router-dom';
 
 import ElnLayout from '@/components/ElnLayout';
 import { EventStatusLabel } from '@/components/EventStatusLabel';
-import { Table as TableQuery } from '@/components/TableQuery';
+import { Table as TableQuery } from '@/components/TableQuery/table';
 import { useTableQuery } from '@/components/TableQuery/hooks/useTableQuery';
 import { Button, Color, Roundness, Variant } from '@/components/tailwind-ui';
 import {
